requirements.txt
streamlit==1.22.0
pinecone-client==2.2.1
sentence-transformers==2.2.2
